% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmeta.R
\name{Deparse}
\alias{Deparse}
\title{Deparsing of expression}
\usage{
Deparse(o)
}
\arguments{
\item{o}{Expression/Object to be deparsed}
}
\value{
single character vector with the deparsed expression

character string with the deparsed R-object
}
\description{
Create single character string from R expression
}
\details{
Calls deparse on argument and pastes together the return value of \code{deparse()} resulting
in a single character string. Operates very similar to \code{dput()}, except, it cannot write to
a file.
}
\examples{
\dontrun{
Deparse(3)
Deparse(1 + 2)
Deparse(matrix(1:10, ncol = 5))
eval(Deparse(matrix(1:10, ncol = 5)))
}
}
\seealso{
[deparse()] which this function wraps

[eval()] for the inverse operation

[dput()] similar function

[dget()] similar to eval of character string
}
