% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\name{DfUniqueByCols}
\alias{DfUniqueByCols}
\title{Reduce data frame to be unique on subset of columns}
\usage{
DfUniqueByCols(d, cols, drop = FALSE)
}
\arguments{
\item{d}{data frame to be made unique}

\item{cols}{columns for which the reduced data frame has to be unique}

\item{drop}{argument passed to subset selection \code{`[`}}
}
\value{
the reduced data frame
}
\description{
Reduce data frame by picking the first row of blocks for which \code{cols} has the same values.
}
