% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsystem.R
\name{file.locate}
\alias{file.locate}
\title{Return absolute path for name searched in search-pathes}
\usage{
file.locate(
  path,
  prefixes = NULL,
  normalize = TRUE,
  as.dirs = TRUE,
  force = FALSE,
  home = TRUE
)
}
\arguments{
\item{path}{path (segment) to be located in standard locations}

\item{prefixes}{prefixes to be prepended to path to check existance}

\item{normalize}{boolean to inidcate whether a normalized path should be returned (absolute)}

\item{as.dirs}{assume that prefixes are pathes, i.e. a slash will be put between path and prefix}

\item{force}{enforces that path and prefix are always joined, otherwise if path is absolute no prefixing is performed}

\item{home}{boolean to indicate whether starting prefix '~' should be interpolated to the home folder}
}
\value{
character vector with path to file or NULL if no file could be located
}
\description{
Search for pathes.
}
