% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\name{inverseOrder_fromOrder}
\alias{inverseOrder_fromOrder}
\title{Calculates inverseOrder, assuming that the argument is already an \code{order}-vector.}
\usage{
inverseOrder_fromOrder(p)
}
\arguments{
\item{p}{obect for which the inverse order is to be calculated}
}
\value{
vector with integers representing the inverse order
}
\description{
Calculates inverseOrder, assuming that the argument is already an \code{order}-vector.
}
