% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstatistic.R
\name{lhPreparePars}
\alias{lhPreparePars}
\title{Build wrapper function around likelihood}
\usage{
lhPreparePars(
  pars,
  defaults = lhSpecificationDefaults$default,
  spec = lhSpecificationDefault,
  template = pars
)
}
\arguments{
\item{template}{parameter specification used as template (usually richest parametrization tb reduced
for other hypotheses)}
}
\description{
Build wrapper function around likelihood
}
