% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\name{order_align}
\alias{order_align}
\title{Return vector that reorders v to equal reference.

Assuming that two arguments are permutaions of each other, return a vector of indeces such that \code{all(reference == v[order_align(reference, v)]) == TRUE} for all vectors \code{reference, v}.}
\usage{
order_align(reference, v)
}
\arguments{
\item{reference}{vector with the reference ordering}

\item{v}{vector that is to be ordered the same way as \code{reference}}
}
\value{
vector of indeces so that \code{v[return_value]} is the same as \code{reference}
}
\description{
Return vector that reorders v to equal reference.

Assuming that two arguments are permutaions of each other, return a vector of indeces such that \code{all(reference == v[order_align(reference, v)]) == TRUE} for all vectors \code{reference, v}.
}
