% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibilitySI.R
\name{plausibility,plausibilityFamily-method}
\alias{plausibility,plausibilityFamily-method}
\title{Compute plausibility}
\usage{
\S4method{plausibility}{plausibilityFamily}(this, optMethod = "grid", Niter = 1L, ...)
}
\arguments{
\item{this}{Plausibility family object}

\item{optMethod}{Optimization method to use. Either 'grid' (default) or 'optim'}

\item{Niter}{Number of iterations used when finding the plausibility estimate}

\item{...}{Arguments passed to the optimizer}
}
\description{
Take a plausibility-family object and compute the plausibility estimate and P-value
}
