% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility.R
\name{PlausibilityBinomialWeighted}
\alias{PlausibilityBinomialWeighted}
\title{Compute weighted plausibility for a model comparison of binomial outcome}
\usage{
PlausibilityBinomialWeighted(
  f0,
  f1,
  data,
  Nsi = 1000L,
  Niter = 2L,
  ...,
  optMethod = "grid",
  Nbinom = 1L
)
}
\arguments{
\item{f0}{Null model given as R formula}

\item{f1}{Alternative model given as R formula}

\item{data}{data frame}

\item{Nsi}{Number of stochastic integration samples (defaults to 1e3L)}

\item{Niter}{Number of iterations used when finding the plausibility estimate}

\item{...}{Arguments passed to the optimizer}

\item{optMethod}{Optimization method to use. Either 'grid' (default) or 'optim'}

\item{Nbinom}{Number of repetions in the binomial outcome}
}
\description{
A model comparison of binomial outcome for nested models is computed. The plausibility estimate and P-value are returned together with a glm-fit.
}
