% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibilityBase.R
\name{PlausibilityRegion}
\alias{PlausibilityRegion}
\title{Compute full or marginal plausibility region}
\usage{
PlausibilityRegion(
  f0,
  f1 = NULL,
  data,
  family = "binomial",
  level = 0.95,
  Nsi = 1000L,
  Niter = 1L,
  Napprox = 20L,
  optMethod = "optim",
  ...,
  plClass = "plausibilityGlm",
  initArgs = list(),
  sigmaScale = 3
)
}
\arguments{
\item{f0}{Null model given as R formula}

\item{f1}{optional formula containing variables to be marginalized over}

\item{data}{data frame}

\item{Nsi}{Number of stochastic integration samples (defaults to 1e3L)}

\item{Niter}{Number of iterations used when finding the plausibility estimate}

\item{optMethod}{Optimization method to use. Either 'grid' (default) or 'optim'}

\item{...}{Arguments passed to the optimizer}
}
\description{
glm-fit.
}
