% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstatistic.R
\name{quantileNormalization}
\alias{quantileNormalization}
\title{Quantile normalization of frame/matrix with respect to reference distribution}
\usage{
quantileNormalization(
  reference,
  data,
  direction = 2,
  impute = TRUE,
  ties = "random",
  center = TRUE,
  referenceDirection = direction
)
}
\arguments{
\item{reference}{numeric vector with realizations from the target distribution}

\item{data}{data frame or matrix with data to be normalized}

\item{direction}{is \code{data} organized per row or column?}
}
\description{
Distribution to be normalized are represented as columns or rows of a matrix/data frame.
Each value is replaced by the quantile of the reference distribution as given by the value of the
empirical distribution function of the given value.
}
\examples{
d = sapply(1:20, rnorm(1e4));
dNorm = quantileNormalization(as.vector(d), d)
}
