% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\name{Rank}
\alias{Rank}
\title{Computes order so that inverseOrder after order is the identity

Caculate ranks for arguemnt \code{p}. Works on vactors and data frames.}
\usage{
Rank(p)
}
\arguments{
\item{p}{object for which ranks are to be comptued}
}
\value{
vector of ranks of elements of \code{p}
}
\description{
Computes order so that inverseOrder after order is the identity

Caculate ranks for arguemnt \code{p}. Works on vactors and data frames.
}
