% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\name{Seq}
\alias{Seq}
\title{Produce constrained sequences}
\usage{
Seq(from, to, ..., ascending = TRUE, descending = !ascending, neg = FALSE)
}
\arguments{
\item{from}{starting value}

\item{to}{ending value}

\item{...}{parameters passed on to \code{seq}}

\item{ascending}{restrict sequences to be ascending; return empty list if to < from}

\item{descending}{restrict sequences to be descending; return empty list if from < to}

\item{neg}{boolean to indicate wheter sequence should be negated before return}
}
\value{
sequence from \code{from} to \code{to}
}
\description{
This is a wrapper around seq that adds constraints. Setting ascending, descending to NA reverts to
standard \code{seq} behaviour.
}
