% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\name{Substr}
\alias{Substr}
\title{Return sub-strings indicated by positions or produce a string by substituting those strings with
replacements}
\usage{
Substr(s, start, length, replacement)
}
\arguments{
\item{s}{template string}

\item{start}{vector of start positions of substrings to substitute}

\item{length}{vector of lengthes of substrings to substitute}

\item{replacement}{vector of strings to subsitute. If missing, \code{Substr} returns sub-strings indicated
by start/length}
}
\value{
character vector containing extracted sub-strings
}
\description{
The function behaves similar to sprintf, except that character sequences to be substituted are
indicated by name.
}
