% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdata.R
\name{valueCombinations}
\alias{valueCombinations}
\title{Return all value combinations appearing in a data frame}
\usage{
valueCombinations(d)
}
\arguments{
\item{d}{data frame for which value combinations are to be caclulated}
}
\value{
list with all value combinations present in \code{d}
}
\description{
Return all value combinations appearing in a data frame
}
